test: TokenRepository 테스트 개선 및 MockHTTPClient 기능 추가
TokenRepository 테스트를 개선하고 MockHTTPClient에 스텁 응답 기능을 추가했습니다.
- MockHTTPClient에 stubResponse 프로퍼티 추가로 정상 응답 시뮬레이션 가능
- TokenRepositoryImplTests 삭제 및 TokenRepositoryTests로 통합
- 테스트 코드 정리 및 개선
